+++ browser/ui/views/brave_actions/brave_actions_container.h
@@ -14,7 +14,7 @@
 #include "base/scoped_observation.h"
 #include "brave/browser/extensions/api/brave_action_api.h"
 #include "brave/browser/ui/views/brave_actions/brave_rewards_action_stub_view.h"
-#include "brave/browser/ui/views/brave_actions/brave_today_action_view.h"
+#include "brave/browser/ui/views/brave_actions/brave_news_action_view.h"
 #include "brave/components/brave_rewards/browser/rewards_service.h"
 #include "chrome/browser/extensions/api/extension_action/extension_action_api.h"
 #include "chrome/browser/ui/browser.h"
