+++ browser/ui/views/brave_actions/brave_news_action_view.h
